+++ src/main/scala/tscfg/generators/Generator.scala
@@ -25,7 +25,8 @@ abstract class Generator(genOpts: GenOpts) {
   */
 case class GenOpts(packageName: String,
                    className: String,
-                   j7: Boolean
+                   j7: Boolean,
+                   useBacksticks: Boolean = false
                   )
 
 case class GenResult(code: String = "?",